split implementation into modules
